# OTP/Integrations/openapi-v1.yaml
openapi: 3.1.0
info:
  title: OTPAuth 2FA API
  version: 1.0.0-draft
  summary: Draft REST API for the first release of the OTPAuth 2FA platform.
  description: |
    API for second-factor challenge orchestration, TOTP enrollment, and mobile device activation.
    This contract is a draft for the MVP release and is expected to evolve before implementation.
  license:
    name: Proprietary
    url: https://otpauth.test/license
servers:
  - url: https://api.otpauth.test
    description: Production placeholder
  - url: https://sandbox.otpauth.test
    description: Sandbox placeholder
...
paths:
...
  /api/v1/enrollments/totp:
    post:
      tags: [Enrollments]
      summary: Start TOTP enrollment
      operationId: startTotpEnrollment
      security:
        - IntegrationOAuth2:
            - enrollments:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartTotpEnrollmentRequest'
      responses:
        '201':
          description: Enrollment started
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpEnrollment'
              examples:
                pending:
                  value:
                    enrollmentId: "a75bf0b8-3a40-418e-a336-0cb7b18ff22a"
                    status: pending
                    secretUri: "otpauth://totp/OTPAuth:ivan.petrov?secret=BASE32SECRET&issuer=OTPAuth"
                    qrCodePayload: "otpauth://totp/OTPAuth:ivan.petrov?secret=BASE32SECRET&issuer=OTPAuth"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/enrollments/totp/{enrollmentId}:
    get:
      tags: [Enrollments]
      summary: Get TOTP enrollment status
      operationId: getTotpEnrollment
      security:
        - IntegrationOAuth2:
            - enrollments:write
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      responses:
        '200':
          description: Enrollment state
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpEnrollment'
              examples:
                confirmed:
                  value:
                    enrollmentId: "a75bf0b8-3a40-418e-a336-0cb7b18ff22a"
                    status: confirmed
                    hasPendingReplacement: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/enrollments/totp/{enrollmentId}/confirm:
    post:
      tags: [Enrollments]
      summary: Confirm TOTP enrollment with a code from the authenticator
      operationId: confirmTotpEnrollment
      security:
        - IntegrationOAuth2:
            - enrollments:write
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmTotpEnrollmentRequest'
      responses:
        '200':
          description: Enrollment confirmed
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpEnrollment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/enrollments/totp/{enrollmentId}/replace:
    post:
      tags: [Enrollments]
      summary: Start safe TOTP enrollment replacement
      operationId: replaceTotpEnrollment
      security:
        - IntegrationOAuth2:
            - enrollments:write
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      responses:
        '200':
          description: Replacement started
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpEnrollment'
              examples:
                replacementPending:
                  value:
                    enrollmentId: "a75bf0b8-3a40-418e-a336-0cb7b18ff22a"
                    status: confirmed
                    hasPendingReplacement: true
                    secretUri: "otpauth://totp/OTPAuth:ivan.petrov?secret=BASE32SECRET&issuer=OTPAuth"
                    qrCodePayload: "otpauth://totp/OTPAuth:ivan.petrov?secret=BASE32SECRET&issuer=OTPAuth"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/enrollments/totp/{enrollmentId}/revoke:
    post:
      tags: [Enrollments]
      summary: Revoke TOTP enrollment
      operationId: revokeTotpEnrollment
      security:
        - IntegrationOAuth2:
            - enrollments:write
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      responses:
        '200':
          description: Enrollment revoked
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpEnrollment'
              examples:
                revoked:
                  value:
                    enrollmentId: "a75bf0b8-3a40-418e-a336-0cb7b18ff22a"
                    status: revoked
                    hasPendingReplacement: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    IntegrationOAuth2:
      type: oauth2
      description: OAuth 2.0 client credentials flow for integration clients.
      flows:
        clientCredentials:
          tokenUrl: https://api.otpauth.test/oauth2/token
          scopes:
            challenges:write: Create challenges and submit second-factor data.
            challenges:read: Read challenge status.
            enrollments:write: Start and confirm factor enrollments.
            devices:write: Activate and revoke mobile devices.
    DeviceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token issued to a registered mobile device.
  parameters:
...
    EnrollmentId:
      name: enrollmentId
      in: path
      required: true
      description: Enrollment identifier.
      schema:
        type: string
        format: uuid
...
  headers:
    X-Request-Id:
      description: Server-generated request identifier.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidPayload:
              value:
                type: "/errors/validation"
                title: "Validation failed"
                status: 400
                detail: "preferredFactors must not be empty"
                traceId: "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-00"
    Unauthorized:
      description: Authentication failed
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Resource state conflict
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        traceId:
          type: string
...
    EnrollmentStatus:
      type: string
      enum: [pending, confirmed, revoked]
...
    StartTotpEnrollmentRequest:
      type: object
      required: [tenantId, externalUserId]
      properties:
        tenantId:
          type: string
          format: uuid
        externalUserId:
          type: string
          minLength: 1
        issuer:
          type: string
        label:
          type: string
    ConfirmTotpEnrollmentRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: '^[0-9]{6}$'
    TotpEnrollment:
      type: object
      required: [enrollmentId, status, hasPendingReplacement]
      properties:
        enrollmentId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/EnrollmentStatus'
        hasPendingReplacement:
          type: boolean
        secretUri:
          type: string
          description: Present only in start/replace responses that return one-time provisioning material.
        qrCodePayload:
          type: string
          description: Present only in start/replace responses; currently equals secretUri and is intended for QR generation.
...